+++ CHANGELOG.md
@@ -7,6 +7,18 @@
   - Support functions as props in `renderOption` and `filter` props
   - Added `clearButtonProps` and `clearable` props
 
+- New Props #609 by @AnnMarieW
+  - `headerControlOrderProp` to calendar components
+  - `presets` to `DatePicker` and  `DateTimePicker`
+  - `autoContrast` to `Tooltip` and `FloatingTooltip`
+  - `domain` to `Slider` and `RangeSlider`
+  - `pushOnOverlap` to `RangeSlider`
+  - `bdrs` new style prop to for `borderRadius`
+  - `getYearControlProps`, `getMonthControlProps`, `getDayProps`, `renderDay` to date components
+
+### Changed
+- Upgraded to latest Mantine (8.1.2)
+
 # 2.0.0
 
 ### Changed